backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
pytest==7.4.3
httpx==0.25.1 